app/requirements-build.txt
# Production

boto3
celery>=5,<6
channels[daphne]<5
channels-redis<5
django>=5,<6
django-celery-beat<3
django-redis>=5,<6
django-storages==1.14
django-cors-headers<5
duo_universal
geopy
gunicorn
httpx
markdown
mysqlclient
packaging
Pillow
psycopg[binary]
python-decouple
python-dotenv
python-magic
pytimeparse2
sentry-sdk
uvicorn[standard]
validators


# Development

#django-debug-toolbar<6
#django-extensions
#watchdog
#Werkzeug


# Tools

#bandit
#black
#coverage
#flake8
#pytest-playwright
#mypy
#ruff
